pyproject.toml
# Build system configuration - tells Python how to package this project
[build-system]
requires = ["setuptools>=68.0", "wheel"]  # Need setuptools and wheel to create distributions
build-backend = "setuptools.build_meta"   # Use setuptools as the build backend

# Core project metadata
[project]
name = "house-of-plantagent"              # Package name (used in pip install)
version = "0.1.0"                         # Semantic versioning: major.minor.patch
description = "A vegan/plant-based nutrition LLM agent with RAG, grounding, and agentic orchestration."  # One-liner summary
requires-python = ">=3.10"                # Minimum Python version required
authors = [{name = "Evelyne Kroval"}]     # Project author(s)
readme = "README.md"                      # Path to README displayed on PyPI
license = {text = "MIT"}                  # License type

# Runtime dependencies - installed when user runs `pip install house-of-plantagent`
dependencies = [
    "requests>=2.31.0",      # HTTP library for web scraping recipes
    "python-dotenv>=1.0.0",  # Loads environment variables from .env files
    "pytest>=7.4.0",         # Testing framework
    "pytest-cov>=4.1.0",     # Code coverage measurement plugin for pytest
    "ipykernel>=7.1.0",
]

# Optional dependencies - installed with `pip install house-of-plantagent[dev]`
[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",           # Fast linter (checks for bugs, style issues)
    "mypy>=1.5.0",           # Static type checker (catches type errors before runtime)
]

# Where setuptools looks for package code
[tool.setuptools.packages.find]
where = ["src"]                           # All Python packages are in src/ directory

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]                     # Run tests from tests/ folder
python_files = ["test_*.py"]              # Only files matching test_*.py are test files
addopts = "-v --cov=src --cov-report=term-only"  # Verbose output + coverage report for src/

# ruff (linter) configuration
[tool.ruff]
line-length = 100                         # Match black's line length
target-version = "py310"                  # Lint for Python 3.10
select = ["E", "F", "W"]                  # E=syntax errors, F=undefined names, W=warnings
